Create MainLayout Component
Creat component MainLayout.js, make the app.js more concise and clear. Now all the layout-related files (including main layout, sider and header) are in the layouts folder.
