+++ client/src/layouts/MainLayout.css
@@ -1,0 +1,5 @@
+@import '~antd/dist/antd.css';
+
+.sider {
+  box-shadow: 2px 0 6px rgba(0, 21, 41, 0.35);
+}
