+++ client/src/pages/Home.css
@@ -1,11 +1,11 @@
-.content{
-    margin: 30px 40px;
+.banner{
+  text-align: center;
 }
 
 .content .banner h1{
     font-size: 48px;
     font-weight: 600;
-    margin: 8px 0 20px;
+    margin-bottom: 20px;
     letter-spacing: 0;
     font-family: Avenir,"Chinese Quote",-apple-system,BlinkMacSystemFont,"Segoe UI","PingFang SC","Hiragino Sans GB","Microsoft YaHei","Helvetica Neue",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
     font-variant: tabular-nums;
