+++ client/src/pages/Login.css
@@ -1,11 +1,4 @@
 .login-form {
-    max-width: 300px;
-    margin-left: auto; 
-    margin-right: auto;
-  }
-.login-form-forgot {
-    float: right;
-  }
-.login-form-button {
-    width: 100%;
+  margin-left: auto;
+  margin-right: auto;
 }
