+++ client/src/pages/Profile.js
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 class Profile extends Component {
 	render(){
 		return (
-			<h1>Profile</h1>
+      <h1>Profile</h1>
 		);
 	};
 }
